Refactor package locations and more
