+++ src/scenarios/NeuralXOR.java
@@ -1,0 +1,11 @@
+package scenarios;
+
+import neuralnetwork.NeuralNetworkManager;
+
+public class NeuralXOR {
+	public static void main(String[] args) {
+		NeuralNetworkManager manager = new NeuralNetworkManager();
+		manager.testXOR();
+
+	}
+}
