--- src/simulation/Test.java
@@ -1,9 +1,0 @@
-package simulation;
-
-public class Test {
-
-	public static void main(String[] args) {
-		
-	}
-
-}
